VFS-2085 Do not communicate with server in event handlers.
If a location/attributes happened to expire before event handler was
called, the handler would communicate with the server to refresh its
caches. In case of an error, an exception would be thrown that wouldn't
be handled (because event handlers do not work in user's context) and
cause std::terminate() to be called.